requirements.txt
django~=5.0.7
django-allauth~=0.63.6
djangorestframework~=3.15.2
django-tailwind~=3.8.0
django-browser-reload~=1.13.0
python-dateutil~=2.9.0.post0
django-environ~=0.11.2